Lesson 13 (#25)
* Lesson 13

* konsultacja 9.08.2023

+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceController.java
@@ -22,7 +22,7 @@ public class InvoiceController implements InvoiceApi {
   }
 
   @Override
-  public int addInvoice(@RequestBody Invoice invoice) {
+  public long addInvoice(@RequestBody Invoice invoice) {
     return invoiceService.save(invoice);
   }
