+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -20,7 +20,7 @@ public class FileBasedDatabase implements Database {
   private final JsonService jsonService;
 
   @Override
-  public int save(Invoice invoice) {
+  public long save(Invoice invoice) {
     try {
       invoice.setId(idService.readLastIdFromFile());
       fileService.appendLineToFile(databasePath, jsonService.convertToJson(invoice));
@@ -32,7 +32,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(long id) {
     try {
       return fileService.readLines(databasePath)
           .stream()
@@ -57,7 +57,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+  public Optional<Invoice> update(long id, Invoice updatedInvoice) {
     try {
       List<String> allInvoices = fileService.readLines(databasePath);
       var invoicesWithoutInvoiceWithGivenId = allInvoices
@@ -80,7 +80,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(long id) {
     try {
       var allInvoices = fileService.readLines(databasePath);
       var invoicesExceptDeleted = allInvoices
@@ -99,7 +99,7 @@ public class FileBasedDatabase implements Database {
     }
   }
 
-  private boolean containsId(String line, int id) {
+  private boolean containsId(String line, long id) {
     return line.contains("{\"id\":" + id + ",\"number\"");
   }
 }
